fix: change generics id type from Int to Long

+++ src/main/java/com/tweteroo/api/repository/UserRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.tweteroo.api.model.TweterooUser;
 
-public interface UserRepository extends JpaRepository<TweterooUser, Integer> {
+public interface UserRepository extends JpaRepository<TweterooUser, Long> {
   public Optional<TweterooUser> findByUsername(String username);
 }
